The length argument of the USB functions for getting/setting HID reports/features should be uint16_t (instead of int). (dm)
git-svn-id: svn://mielke.cc/main/brltty@5254 91a5dbb7-01b9-0310-9b5f-b28072856b6e

+++ Programs/io_usb.h
@@ -507,7 +507,7 @@ extern int usbHidGetReport (
   unsigned char interface,
   unsigned char report,
   void *buffer,
-  int length,
+  uint16_t length,
   int timeout
 );
 
@@ -516,7 +516,7 @@ extern int usbHidSetReport (
   unsigned char interface,
   unsigned char report,
   const void *buffer,
-  int length,
+  uint16_t length,
   int timeout
 );
 
@@ -525,7 +525,7 @@ extern int usbHidGetFeature (
   unsigned char interface,
   unsigned char report,
   void *buffer,
-  int length,
+  uint16_t length,
   int timeout
 );
 
@@ -534,7 +534,7 @@ extern int usbHidSetFeature (
   unsigned char interface,
   unsigned char report,
   const void *buffer,
-  int length,
+  uint16_t length,
   int timeout
 );
